% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse.R
\name{getResources}
\alias{getResources}
\title{Get Resource From Downloaded Table}
\usage{
getResources(tb, index)
}
\arguments{
\item{tb}{(data.table). The output of \code{\link{dataSetMetadataSearch}} or \code{\link{resourceMetadataSearch}}.}

\item{index}{(integer). The index of target.}
}
\value{
A data.frame. The resources of \code{index}th row in \code{tb} will be returned.
}
\description{
Get Resource From Downloaded Table
}

