% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resourceAquire.R
\name{resourceAquire}
\alias{resourceAquire}
\title{resourceAquire}
\usage{
resourceAquire(resourceId)
}
\arguments{
\item{resourceId}{which is one column of query result by using getResources(). Please read seealso to get more information.}
}
\description{
resourceAquire
}
\references{
2.3 of \url{http://data.taipei/opendata/developer;jsessionid=4197BA531430955C04015CD2F2CF3699}
}
\seealso{
getResources()
}

