% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resourceMetadataSearch.R
\name{resourceMetadataSearch}
\alias{resourceMetadataSearch}
\title{resourceMetadataSearch}
\usage{
resourceMetadataSearch(q = NULL, limit = 1000, offset = 0)
}
\description{
resourceMetadataSearch
}

