% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSetMetadataSearch.R
\name{dataSetMetadataSearch}
\alias{dataSetMetadataSearch}
\title{Search Dataset of \url{http://data.taipei} via Metadata}
\usage{
dataSetMetadataSearch(q = NULL, limit = 1000, offset = 0)
}
\arguments{
\item{q}{(character). Keyword of the datasets. Please read the details for more information.}

\item{limit}{(integer). Limit the number of returned records.}

\item{offset}{(integer). The start position of the returned records.  Please read the details for more information.}
}
\value{
A data.table. A table of related datasets.
}
\description{
Search Dataset of \url{http://data.taipei} via Metadata
}
\details{
You could directly put the keyword into \code{q}. For example, \code{dataSetMetadataSearch(q = "youbike")}.
 \code{q} also accept advanced usage such as \code{q="title:youbike cost:free"} which will return the records whose title contains "youbike" and cost contains "free". The schema is "field-1:condition-1 field-2:condition-2 ...". The blank space separates each condition. For more details, please check the reference.

 For {offset}, here is an example. If \code{offset = 5}, the \code{dataSetMetadataSearch} will return the \code{1 + 5 = 6}th row.
}
\examples{
\dontrun{
rs <- dataSetMetadataSearch(q = "youbike")
rid <- getResources(rs, 1)
suppressWarnings(df <- resourceAquire(rid$resourceId[1]))
}
}
\references{
\url{http://tpeodck.gitbooks.io/data-taipei-developer-guide/content/chapter2section1.html}
}
\seealso{
\code{\link{getResources}}
}

