% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSetMetadataSearch.R
\name{dataSetMetadataSearch}
\alias{dataSetMetadataSearch}
\title{dataSetMetadataSearch}
\usage{
dataSetMetadataSearch(q = NULL, limit = 1000, offset = 0)
}
\arguments{
\item{q}{(character). Keyword of datasets. Read the details for more information.}

\item{limit}{(integer). Limit quanty of returned records to set a proper scale of data output.}

\item{offset}{(integer). The start position to show of query output.  read the details for more information.}
}
\description{
dataSetMetadataSearch
}
\details{
This is the example of q advanced usage. Such as q="title:youbike cost:free", which will return the result's title contains "youbike" and cost contains "free". The method is "condition-1 condition-2 ...", which uses one blank space to separate each condition, and you can input one or more conditions to query.

Please read the references for more information.

  This is the example of offset usage. Such as offset = 5, which will return the 1 + 5 = 6th dataset.

  Please read the references for more information.
}
\references{
\url{http://tpeodck.gitbooks.io/data-taipei-developer-guide/content/chapter2section1.html}
}

