% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resourceAquire.R
\name{resourceAquire}
\alias{resourceAquire}
\title{resourceAquire}
\usage{
resourceAquire(rid)
}
\arguments{
\item{resourceId}{which is one column of query result by using getResources(). Please read seealso to get more information.}
}
\description{
resourceAquire
}
\seealso{
getResources()
}

