% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resourceAquire.R
\name{resourceAquire}
\alias{resourceAquire}
\title{Download the Resource from \url{http://data.taipei}}
\usage{
resourceAquire(resourceId)
}
\arguments{
\item{resourceId}{(character). The id of the resource.}
}
\value{
A data.frame. The downloaded table of the resource.
}
\description{
Download the Resource from \url{http://data.taipei}
}
\examples{
\dontrun{
rs <- dataSetMetadataSearch(q = "youbike")
rid <- getResources(rs, 1)
suppressWarnings(df <- resourceAquire(rid$resourceId[1]))
}
}
\references{
\url{http://tpeodck.gitbooks.io/data-taipei-developer-guide/content/chapter2section3.html}
}
\seealso{
\code{\link{getResources}}
}

