% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resourceAquire.R
\name{resourceAquire}
\alias{resourceAquire}
\title{resourceAquire}
\usage{
resourceAquire(resourceId)
}
\arguments{
\item{resourceId}{which is one column of query result by using getResources(). Please read seealso to get more information.}
}
\description{
resourceAquire
}
\references{
\url{http://tpeodck.gitbooks.io/data-taipei-developer-guide/content/chapter2section3.html}
}
\seealso{
getResources()
}

