% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resourceMetadataSearch.R
\name{resourceMetadataSearch}
\alias{resourceMetadataSearch}
\title{resourceMetadataSearch}
\usage{
resourceMetadataSearch(q = NULL, limit = 1000, offset = 0)
}
\arguments{
\item{q}{(character). Keyword of datasets, what you want to know. Please see the details for more information.}

\item{limit}{(integer). Limit of returned records, if you just need the part of results.}

\item{offset}{(integer). Distance of the starting point. Please see the details for more information.}
}
\description{
resourceMetadataSearch
}
\details{
This is the example of q advanced usage. Such as q=title:youbike cost:free, which will return the result's title contains "youbike" and cost contains "free". The method is "condition-1 condition-2 ...", which uses one blank space to separate each conditions, and you can input two and more conditions to query. Please see the references for more information.
}
\references{
\url{http://tpeodck.gitbooks.io/data-taipei-developer-guide/content/chapter2section2.html}
}

